+++ pystiche/image/transforms/__init__.py
@@ -1,2 +1,1 @@
 from .transforms import *
-from .processing import *
