+++ pystiche/image/transforms/transforms/__init__.py
@@ -1,0 +1,5 @@
+from .core import *
+from .misc import *
+from .resize import *
+from .motif import *
+from .color import *
